Improve command line interface. Fix --reset-config.
The client now can take an arbitrary number of server string as arguments.
For instance `pypentago 127.0.0.1 serv1.unixpod.com:27321` to connect
to 127.0.0.1 on the default port (26500) and to serv1.unixpod.com on port
27321.

The server now deletes the pidfile when it shuts down.

verbosity is removed from both the server and the client configuration file.
--reset-config is fixed for pypentago and pypentagod.